ssz: Move stateutil.SliceRoot to ssz package
