+++ encoding/ssz/BUILD.bazel
@@ -7,6 +7,7 @@ go_library(
         "helpers.go",
         "htrutils.go",
         "merkleize.go",
+        "slice_root.go",
     ],
     importpath = "github.com/prysmaticlabs/prysm/v5/encoding/ssz",
     visibility = ["//visibility:public"],
